requirements.txt
speedtest-cli
requests
bs4
